services/gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.3
pydantic-settings==2.7.0
redis==5.2.1
